根因与修复 根因：JwtHelper.java:49 中 expiresInMinutes * 60 是 int * int，当 expiresInMinutes = 2147483647 时溢出为 -60，导致 token 过期时间被设为 当前时间 - 60秒，即 token 生成即过期。
// 溢出前：2147483647 * 60 = int 溢出 → -60
Date expiration = new Date(now.getTime() + expiresInMinutes * 60 * 1000L);

// 修复：先将 expiresInMinutes 转为 long
Date expiration = new Date(now.getTime() + (long) expiresInMinutes * 60 * 1000L);
重启应用后即可生效。

+++ src/main/java/com/tonghui/erp/Common/utils/JwtHelper.java
@@ -46,7 +46,7 @@ public class JwtHelper {
         
         // 创建用户声明
         Date now = new Date();
-        Date expiration = new Date(now.getTime() + expiresInMinutes * 60 * 1000L);
+        Date expiration = new Date(now.getTime() + (long) expiresInMinutes * 60 * 1000L);
         
         JwtBuilder builder = Jwts.builder()
                 .claim("userId", userId)
